+++ picoquic/picoquic_internal.h
@@ -300,7 +300,7 @@ extern "C" {
         int peer_addr_len;
         struct sockaddr_storage dest_addr;
         int dest_addr_len;
-        int if_index_dest;
+        unsigned long if_index_dest;
 
 		/* connection state, ID, etc. Todo: allow for multiple cnxid */
 		picoquic_state_enum cnx_state;
